Create Updating Display Test
This test code sets up the display in 6 sections and adds the test number 45 to each section. This code was used to adjust the title and number sizes and locations on the display and to begin writing the update functions that will be used in the final iteration of the system.